requirements.txt
altair==5.2.0
appnope==0.1.3
asttokens==2.4.0
attrs==23.2.0
backcall==0.2.0
black==23.7.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.4.2
executing==2.0.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kaleido==0.2.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mypy==1.6.1
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pipenv==2023.11.15
platformdirs==3.11.0
plotly==5.18.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pyzmq==25.1.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
seaborn==0.13.2
setuptools==69.0.2
six==1.16.0
smmap==5.0.1
SqliteCloud==0.0.30
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.31.0
streamlit-option-menu==0.3.12
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
validators==0.22.0
vanna==0.0.38
virtualenv==20.24.7
wcwidth==0.2.8
Werkzeug==3.0.1
XlsxWriter==3.1.9
zipp==3.17.0
